[FEATURE] Active directory authentication
A custom implementation of the Spring authentication provider for
active directory servers is added.

The security configuration is enhanced so that environment variables
can be used to switch to Active Directory authentication.

A custom implementation of the Active Directory authentication
provider is used because the ActiveDirectoryLdapAuthenticationProvider
that comes with the spring framework is marked final and can not be
extended, see: http://goo.gl/qNBSNo.