Remove nexus.synyx.de repository from POM
Since the referenced repository is not available for anyone outside
of our company (protected via password), it leads to non-successful
builds, dependending on the Maven version in question.

Since there are no actual artifacts, that need to be fetched from
there, we can simply remove it. We keep the distribution management
repositories though, since we use those for automatic deployment and
they don't affect people outside of synyx, who are building from
source.