make last comment directly accessible for ApplicationDetails
For some cases UI rendering depends not on the state of the
application but on the state of the lattest comment. As I don't want to
fiddle around too much with collections in JSP code I make the latest
comment accessible as a model property. If someone has a better
solution - feel free to change this. :)