Use Spring-Boot-Starter-Mail / Replace mock_javamail with greenmail
We want the latest javax mail dependencies but mock_javamail will not ne
updated so we have to find something new, because it does not work with
the latest java mail versions. Greenmail is our choice. We do not have
to change a lot. Just some minor changes and the project is well
documented and active.

With this chance we introduce spring-boot-mail-starter and greenmail as
test mail server