introduce AbsenceService getOpenAbsences for date range
this patch introduces new datastructures and options to get absences from.
overviews won't be fixed with this... The previous behaviour should still be the same.
the fix will follow with next commits.

* `AbsencePeriod` will replace `Absence` soon. introduce the new datastructure to don't break calendar stuff.
* add explicit datastructures (e.b. Sick and Vacation Absences)
  to make it easier to reason about Sick Notes, Vacation (or soon: custom absence types)
  -> persons with role USER may not allowed to see SICK/VACATION but only ABSENT
  typing helps here, hopefully

AbsencePeriod consists of
- morning absence
- noon absence
- or both; while morning could be sick and noon vacation in theory

public holidays are not considered as absence.
someone could work despite the public holiday (our app does not support this currently)

public holidays have to be handled by the consumer.
see `AbsenceApiController` or `AbsenceOverviewViewController`