Calculate vacation days left after april only if requested
An application must be checked to ensure enough vacation days left.

The calculation of requested vacation day to fulfill the application is
based on workdays which need to be satisfied by vacation days.

workdays before april can be satisfied by vacation days or remaining
vacation days of the last year.
workdays after april can be satisfied by vacation days or remaining
vacation days of the last year not expiring.

Following rule was added in this commit:

If an application takes place before april there is no need to check
after april for this application