Make Selenium tests work on Windows
The Selenium integration tests failed when run on Windows. The reason
for the failur is not 100% clear but most likely deals with a limited /dev/shm space
inside the docker container (like described in
https://bugs.chromium.org/p/chromium/issues/detail?id=715363). The
commit fixes that by adding --disable-dev-shm-usage to the ChromeOptions
that are used to create the WebDriver.