add all en locales to datepicker
due-date-picker uses Intl.DateTimeFormat to format the date object to a string which is used for screen readers (e.g. "18. März", "18 March" or "March 18"). we could to set the duet-date-picker locale to "de" for instance, however, this would print the month text in german always.

therefore just add all en locales and consider the correct locale when we parse the formatted string back to a date. (to find info how the day should be rendered)