avoid table overflows the available width
'approval-process' should not affect column width which stretches the table width -> overflows the available container width (maybe because of bootstraps float layout? not sure.)

german translation is 'genehmigungsprozess' which is a long word 😅

drawback: word could overflow the column on the left. however, this happens only on a few pixels right now before the layout switches to "cards"