Added property to customize the generated url for calendar sharing
If the server running the application is located behind a reverse proxy, the automatically generated url for the calendar sharing could be not reachable from outside, e.g. when the server runs on a non-standard TCP port but the proxy listens to 80 or 443 ports.

The following property can now be used in the application.properties file to override the standard automatic generation of the url:

uv.calendar.url=http://external.url.of.urlaubsverwaltung/

if the property is missing in the file, the standard logic will be used to generate the url