ignore turbo-frame header and always render whole page
otherwise only the snippet is rendered after user session timeout and user did a new login.
happens when the page is open in the browser, the session timed out and the user searches for a person for example.

we could check if the referer is `login` and then render the full page. however, this seems to be too complex...
so for the moment just remove turbo knowledge from the controllers.