introduce checkbox-switch to enable/disable local vacation expiration for user
drawback:
POSTing the form without javascript results in validation error because "null" is sent for `expiryDate` and `remainingVacationDaysNotExpiring`.
The user did see the actual values in `disabled` input fields, and now has to enter the values. This is kind of confusing...
(with JavaScript the input fields are enabled again and the values are send to the server. so everyting is fine in this case)

we could check the existing account in the account-validator, however.
but I don't want to implement this currently :o)
And JavaScript should be the 99% case. It works without JavaScript... so... it's fine.