Improve avatar rendering for unknown (deleted) person
Open issue: url is set to 'unknown' this causes a bad request to
baseUrl+"deleted" which leads to 404 and afterwards the fallback is
used. Not sure how to replace this without a failing call.

Setting url to '' or null causes a request to baseUrl which works and
does not cause fallback to avatar controller