Add group and resource access claim mapper
If a mapper is enabled the role/group "urlaubsverwaltung_user" must be
active or the user role with a given role prefix. If not the user cannot
be logged in. Jwt and OidcUser claims will be searched for groups or
resource access and th authorities will be merged with the ones from the
database.