Enable partial output while procession in thymeleaf...
... if there is a site with forms and therefore a session is needed it
is not possible to enable partial procession if session is not always
true.
But now we do not have any "free" site with a form. We had one with the
login page via the default login, but that one is gone and therefore we
can use partial output. If there would be any other page with a form or
something else, we would need to enable sessions 'always' for the given
site.