use short dateFormat for datepicker
rendering double digits prohibits input of e.g. `25.07.2024`

given rendered value of `21.07.2024`, selecting `21` and entering a `2`
will rerender the whole input because of a new valid date (duet-date-picker listens to `input`)

with this shorter date format this does not happen anymore.
the input value already is `2.7.2024` after entering `2` and the element won't be rerenderd...
The cursor stays at the correct position and the next value `5` can be entered.

note that the previous issue can still occur...
when the user has entered a double digit month `07` vs `7`.
however, this should happen less than entering days...